Generate the Eclipse classpath from the present jars
Allows changing the manifest without having to update the classpath_base too
Also avoids writing the project or classpath files if they already exist

+++ scripts/eclipse_setup.py
@@ -1,17 +1,24 @@
 import os
+import glob
 
-file_data = ""
-with open("./data/classpath_base.txt", 'r') as file:
-    file_data = file.read()
+if not os.path.exists("./.classpath"):
+    with open("./data/classpath_base.txt", 'r') as file:
+        file_data = file.read()
 
-with open(".classpath", 'w+') as file:
-    file.write(file_data)
+    libs = [os.path.normpath(jar).replace("\\", "/") for jar in glob.glob("./libs/*.jar")
+             if os.path.basename(jar) != "client-mapped.jar"]
+    libs = "\n".join(map(lambda jar: "\t<classpathentry kind=\"lib\" path=\"" + jar + "\"/>", libs))
+    file_data = file_data.replace("**LIB CLASSPATH**", libs, 1)
 
-with open("./data/project_base.txt", 'r') as file:
-    file_data = file.read()
+    with open(".classpath", 'w+') as file:
+        file.write(file_data)
 
-with open(".project", 'w+') as file:
-    file.write(file_data)
+if not os.path.exists("./.project"):
+    with open("./data/project_base.txt", 'r') as file:
+        file_data = file.read()
+
+    with open(".project", 'w+') as file:
+        file.write(file_data)
 
 if not os.path.exists("./bin"):
     os.makedirs("./bin")
